# backend_school_crm/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@school-crm.example
    name: School CRM Support
  description: Residual monolith — auth, subscriptions, Click.uz/Telegram payments,
    messaging, reports, dev panel.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: School CRM API
  version: "1.0"
paths:
  /attendance:
    get:
      description: Returns attendance records for a class and date.
      parameters:
      - description: Class ID
        in: query
        name: classId
        required: true
        type: string
      - description: YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get attendance
      tags:
      - attendance
    post:
      consumes:
      - application/json
      description: Creates or updates attendance records for a class session.
      parameters:
      - description: class_id, date, records (student_id, status)
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Record attendance
      tags:
      - attendance
...
  /classes:
    get:
      description: Returns all classes for a branch.
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List classes
      tags:
      - classes
    post:
      consumes:
      - application/json
      parameters:
      - description: name, branch_id, teacher_id, schedule
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: created class
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create class
      tags:
      - classes
  /classes/{id}:
    delete:
      parameters:
      - description: Class UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: class deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete class
      tags:
      - classes
...
  /students:
    get:
      description: Returns a paginated list of students for the given branch.
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10, max 1000)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: data, total, page, limit, totalPages
          schema:
            additionalProperties: true
            type: object
        "400":
          description: branchId required
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List students
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Creates a new student in the branch.
      parameters:
      - description: Student data (full_name, monthly_payment, status, branch_id)
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: created student
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "402":
          description: subscription_limit_reached
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create student
      tags:
      - students
  /students/{id}:
    delete:
      description: Permanently removes a student record.
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: student deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete student
      tags:
      - students
    get:
      description: Returns a single student by ID.
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: student object
          schema:
            additionalProperties: true
            type: object
        "404":
          description: not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get student
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Partially updates a student record (any field provided in body).
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: updated student
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update student
      tags:
      - students
...
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"